+++ route/main_error_404.py
@@ -1,8 +1,6 @@
 from .tool.func import *
 
-def main_error_404_2(conn):
-    curs = conn.cursor()
-
+def main_error_404(e):
     if os.path.exists('404.html') and flask.request.path != '/':
         return open('404.html', encoding='utf8').read()
     else:
