+++ route/main_tool_other.py
@@ -1,8 +1,6 @@
 from .tool.func import *
 
-def main_other_2(conn):
-    curs = conn.cursor()
-
+def main_tool_other():
     return easy_minify(flask.render_template(skin_check(),
         imp = [load_lang('other_tool'), wiki_set(), wiki_custom(), wiki_css([0, 0])],
         data = '''
