+++ route/main_view_file.py
@@ -1,9 +1,7 @@
 from .tool.func import *
-from . import main_error_404
+from .main_error_404 import main_error_404
 
-def main_file_2(conn, data):
-    curs = conn.cursor()
-
+def main_view_file(data = ''):
     if data == 'robots.txt' and not os.path.exists('robots.txt'):
         return flask.Response('User-agent: *\nDisallow: /\nAllow: /$\nAllow: /w/', mimetype = 'text/plain')
     elif os.path.exists(data):
@@ -12,4 +10,4 @@ def main_file_2(conn, data):
         else:
             return flask.send_from_directory('./', data, mimetype = 'text/xml')
 
-    return main_error_404.main_error_404_2(conn)
+    return main_error_404()
