+++ route/main_view_image.py
@@ -1,9 +1,7 @@
 from .tool.func import *
-from . import main_error_404
+from .main_error_404 import main_error_404
 
-def main_image_view_2(conn, name):
-    curs = conn.cursor()
-
+def main_view_image(name = ''):
     mime_type = re.search(r'([^.]+)$', name)
     if mime_type:
         mime_type = mime_type.group(1).lower()
@@ -15,4 +13,4 @@ def main_image_view_2(conn, name):
             mimetype = 'image/' + mime_type
         )
     else:
-        return main_error_404.main_error_404_2(conn)
+        return main_error_404()
